saga_action_func: simplify types and constraints
This was intended to be a non-breaking change - I'd appreciate any info
to the contrary!

- Introduced ActionFn trait to wrap up the repeated stack of type
  constraints around async functions. (Better name suggestions welcome,
  would have called it ActionFunc, but ActionFunc is already a thing
  that is actually *two* funcs).

- With the Future and Output types now associated on the trait,
  eliminated the type parameters that used to model them.

- Removed unnecessary constraints (e.g. on the Debug impl) which were
  boilerplatey and open to rot.

I was tempted to make ActionFunc private, since users can't call any
operations on it or create an instance that isn't behind a Box dyn, but
I'm concerned that I don't understand why it's pub and I'm worried that
would be a breaking change.