Subsagas! -
Unfortunately the instance id strategy doesn't provide enough information
for the outer saga to identify both its own saga and the nested subsagas it needs :(

We're going to need some other mechanism for this. It may be as simple as
adding another intermediate node with metadata to inform the following nodes
what to do. Or maybe instance_id should instead become a list of key inputs
required  by the following nodes?

Need to think more about this.