Distinguish top-level Saga Dags from subsaga Dags
Mechanically, this change looks like this:

- rename `Dag` -> `SagaDag`
- create new `Dag` type that's agnostic about whether it's going to be used in a
  top-level saga or subsaga.  You can make a `SagaDag` out of it by providing
  parameters (which are always known at DAG construction time for top-level
  sagas).  Or you can make a subsaga out of it using `UserNode::subsaga()` and
  append that to another `DagBuilder` (just like you could before this change).

Why bother?  There were some issues with this before:

- `Dag` always contained saga parameters.  But `Dag` was used to represent
  top-level sagas as well as subsagas.  Subsagas don't have parameters when the
  Dag is constructed.  So if you were building a subsaga, you would build a Dag
  and include some dummy parameter value that was always ignored.
- `Dag` always had a Start and End node.  That sounds reasonable but it doesn't
  really make sense: if this was being used as a subsaga, we always ignored the
  Start and End nodes and replaced them with SubsagaStart and SubsagaEnd
  nodes.  This was more in the way than it sounds because a proper Start node
  has to include saga parameters, and again, we don't know those for a subsaga
  when we're making the Dag.

The net result here is more types.  But I think it better reflects reality.
You've got the `Dag` that's agnostic about whether it's top-level or subsaga.
And then you either add params and make it a top-level saga or you turn it into
a subsaga.

It might be even cleaner to have separate Node types for `Dag` and `SagaDag`.
To match what we do now, `Dag` would have `Action`, `Constant`, `SubsagaStart`,
and `SubsagaDone` (but no `Start` or `End`).  Or we could add `Start` (with no
params) and `End` -- it wouldn't really matter.  `SagaDag` would have the same
kinds of nodes except that it definitely needs `Start` (_with_ params) and
`End`.  This seems excessively pedantic to me right now, especially with all the
other similar types we've got here.