simplify saga_params_for() and make_ancestor_tree()
When loading a saga (either for the first time or during recovery), we
now walk the nodes in order and construct a mapping from each node index
to the node index of the containing saga or subsaga.  This is reasonably
straightforward to do and it makes saga_params_for() and
make_ancestor_tree() much simpler.  Instead of these functions starting
from an arbitrary node, walking towards the root, keeping track of
subsaga depth, ignoring ranges, etc., saga_params_for() can immediately
get to the node containing the information needed and
make_ancestor_tree() can quickly skip over a whole subsaga if it finds
one.