more make_ancestor_tree_node() cleanup
We're now getting the value of a constant node from its output rather
than its config.  It doesn't really matter but this simplifies the code
(and avoids a clone I'd wanted to remove anyway).  We also now verify an
invariant that we don't encounter a duplicate node name in the saga.  We
should validate this in DagBuilder but we do not yet today.