Dynamic sagas + working subsagas (#29)
This code is a substantial change of the existing behavior. DAGs for a
given saga are no longer statically defined at build time through the use of
`SagaTemplateBuilder`s. Instead, DAGs can be dynamically constructed at runtime
through a `DagBuilder`, which enables DAGs of different shapes for a given Saga
operation depending upon user input.

Additionally, the implementation of subsagas has changed. Subsagas are no
longer defined by templates, and launched as separate sagas by a node of the
parent saga. This was challenging to make idempotent, and in its current state
was unsound. Instead, subsagas are now added directly as nodes into the dynamic
DAG via the `DagBuilder`. Subsagas themselves are constructed as `Dag`s and can
get added with other `Node`s via `DagBuilder::append` and
`DagBuilder::append_parallel`. Subsaga parameters come from a parameter node
that is output from the parent saga. When the top level saga is done being constructed
it is packaged up into a `SagaDag`.

In order to enable fully dynamic sagas and subsagas, an `ActionRegistry` was
created, where all actions across all sagas are registered. `Dag`s refer to
these actions by `ActionName`. This allows a `SagaDag` to be serialized,
and when deserialized, run arbitrary Rust action code without having to couple
the structure of the DAG to that rust code as in the prior template driven
design.

There are many other goodies sprinkled throughout, including better tests and
validation. See the comments on PR #29 for details.